+++ CHANGES.md
@@ -5,6 +5,10 @@
   to be updated to use the `DocumentRevision` class. See the updated
   [CRUD guide](https://github.com/cloudant/sync-android/blob/master/doc/crud.md)
   for examples of how to use this class.
+- [BREAKING CHANGE] Index type is now defined as the IndexType enum this effects
+ the following APIs:
+  - Index#getInstance
+  - IndexManger#ensureIndexed
 
 # 0.15.5 (2016-02-25)
 - [FIXED] Issue where `java.lang.RuntimeException: Offer timed out` could be thrown 5 minutes after